requirements.txt
Flask==3.0.2
huggingface-hub==0.22.2
Jinja2==3.1.3
pillow==10.2.0
pyOpenSSL==24.1.0
torch==2.2.2
torchvision==0.17.2



